Tweak for OS/2 build happiness.

+++ miscutil.c
@@ -36,6 +36,9 @@ const char miscutil_rcs[] = "$Id$";
  *
  * Revisions   :
  *    $Log$
+ *    Revision 1.41  2006/08/18 16:03:17  david__schmidt
+ *    Tweak for OS/2 build happiness.
+ *
  *    Revision 1.40  2006/08/17 17:15:10  fabiankeil
  *    - Back to timegm() using GnuPG's replacement if necessary.
  *      Using mktime() and localtime() could add a on hour offset if
@@ -1062,7 +1065,7 @@ long int pick_from_range(long int range)
 #ifndef HAVE_RANDOM
    unsigned int weak_seed;
 
-   weak_seed = (unsigned int)(time(NULL) | range);
+   weak_seed = (unsigned int)((unsigned int)time(NULL) | (unsigned int)range);
    srand(weak_seed);
    /*
     * Some rand implementations aren't that random and return mostly
